+++ debugreq/debugreq-details.html
@@ -1,7 +1,7 @@
 <html>
   <head>
     <title>Debug Request Details</title>
-    <link type="text/css" rel="stylesheet" href="css/materialize.css" media="screen,projection"/>
+    <link type="text/css" rel="stylesheet" href="/css/materialize.css" media="screen,projection"/>
   </head>
   <body>
     <div class="container" style="margin-top: 100px;">
@@ -33,7 +33,7 @@
                   <div id="no-info-msg" style="display: none;"><p>No information available.</div>
                 </div>
                 <div class="col s5 right-align">
-                  <a id="closeButton" class="waves-effect waves-light btn">Close Tab</a>
+                  <a id="closeButton" class="light-blue waves-effect waves-light btn">Close Tab</a>
                 </div>
               </div>
             </div>
@@ -42,10 +42,10 @@
 
     </div> <!-- container -->
 
-    <script type="text/javascript" src="js/jquery-3.1.1.min.js"></script>
-    <script type="text/javascript" src="js/materialize.min.js"></script>
-    <script type="text/javascript" src="js/common.js"></script>
-    <script type="text/javascript" src="js/debugdetails.js"></script>
+    <script type="text/javascript" src="/js/jquery-3.1.1.min.js"></script>
+    <script type="text/javascript" src="/js/materialize.min.js"></script>
+    <script type="text/javascript" src="/js/common.js"></script>
+    <script type="text/javascript" src="/debugreq/debugreq-details.js"></script>
   </body>
 </html>
