+++ debugreq/debugreq-details.js
@@ -1,9 +1,5 @@
-//Ricky Yu owns this page, please do comment out the section you edit or added so he is aware of the changes
-
 $(document).ready(function(){
-  chrome.runtime.getBackgroundPage(function (backgroundpage){
-    backgroundpage._gaq.push(['_trackEvent', 'Debug_details_page', 'loaded']);
-   });
+  chrome.runtime.sendMessage({type: "gaq", target: "Debug_details_page", behavior: "loaded"});
 
     var passedId = getUrlParameter('id');
   
@@ -148,9 +144,7 @@ $(document).ready(function(){
     });
   
     $('#closeButton').click(function(){ 
-      chrome.runtime.getBackgroundPage(function (backgroundpage){
-        backgroundpage._gaq.push(['_trackEvent', 'Debug_details_page_closebtn', 'clicked']);
-       });
+      chrome.runtime.sendMessage({type: "gaq", target: "Debug_details_page_closebtn", behavior: "clicked"});
       closeCurrentTab(); 
     });
