+++ debugreq/debugreq-history.html
@@ -1,7 +1,7 @@
 <html>
   <head>
     <title>Debug History Details</title>
-    <link type="text/css" rel="stylesheet" href="css/materialize.css" media="screen,projection"/>
+    <link type="text/css" rel="stylesheet" href="/css/materialize.css" media="screen,projection"/>
   </head>
   <body>
     <div class="container" style="margin-top: 100px;">
@@ -52,10 +52,10 @@
 
     </div> <!-- container -->
 
-    <script type="text/javascript" src="js/jquery-3.1.1.min.js"></script>
-    <script type="text/javascript" src="js/materialize.min.js"></script>
-    <script type="text/javascript" src="js/common.js"></script>
-    <script type="text/javascript" src="js/debug-history.js"></script>
+    <script type="text/javascript" src="/js/jquery-3.1.1.min.js"></script>
+    <script type="text/javascript" src="/js/materialize.min.js"></script>
+    <script type="text/javascript" src="/js/common.js"></script>
+    <script type="text/javascript" src="/debugreq/debugreq-history.js"></script>
   </body>
 </html>
