+++ fastpurge/fastpurge-details.html
@@ -1,7 +1,7 @@
 <html>
   <head>
     <title>Purge Request Details</title>
-    <link type="text/css" rel="stylesheet" href="css/materialize.css" media="screen,projection"/>
+    <link type="text/css" rel="stylesheet" href="/css/materialize.css" media="screen,projection"/>
   </head>
   <body>
     <div class="container" style="margin-top: 100px;">
@@ -36,7 +36,7 @@
                   <div id="no-info-msg" style="display: none;"><p>No information available.</div>
                 </div>
                 <div class="col s5 right-align">
-                  <a id="closeButton" class="waves-effect waves-light btn">Close Tab</a>
+                  <a id="closeButton" class="btn light-blue waves-effect waves-light">Close Tab</a>
                 </div>
               </div>
             </div>
@@ -45,10 +45,10 @@
 
     </div> <!-- container -->
 
-    <script type="text/javascript" src="js/jquery-3.1.1.min.js"></script>
-    <script type="text/javascript" src="js/materialize.min.js"></script>
-    <script type="text/javascript" src="js/common.js"></script>
-    <script type="text/javascript" src="js/purgedetails.js"></script>
+    <script type="text/javascript" src="/js/jquery-3.1.1.min.js"></script>
+    <script type="text/javascript" src="/js/materialize.min.js"></script>
+    <script type="text/javascript" src="/js/common.js"></script>
+    <script type="text/javascript" src="/fastpurge/fastpurge-details.js"></script>
   </body>
 </html>
