+++ fastpurge/fastpurge-history.html
@@ -1,8 +1,8 @@
 <html>
   <head>
     <title>Purge History</title>
-    <link type="text/css" rel="stylesheet" href="css/materialize.css" media="screen,projection"/>
-    <link type="text/css" rel="stylesheet" href="css/common.css" media="screen,projection"/>
+    <link type="text/css" rel="stylesheet" href="/css/materialize.css" media="screen,projection"/>
+    <link type="text/css" rel="stylesheet" href="/css/common.css" media="screen,projection"/>
   </head>
   <body>
     <div class="container" style="margin-top: 100px;">
@@ -20,7 +20,7 @@
 									<thead>
 										<tr>
 											<th>Time(GMT)</th>
-											<th>Token</th>
+											<th>Credential</th>
 											<th>UpdateType</th>
 											<th>PurgeType</th>
 											<th>PurgeId</th>
@@ -54,10 +54,10 @@
       </div>
     </div> <!-- container -->
 
-    <script type="text/javascript" src="js/jquery-3.1.1.min.js"></script>
-    <script type="text/javascript" src="js/materialize.min.js"></script>
-    <script type="text/javascript" src="js/common.js"></script>
-    <script type="text/javascript" src="js/purge-history.js"></script>
+    <script type="text/javascript" src="/js/jquery-3.1.1.min.js"></script>
+    <script type="text/javascript" src="/js/materialize.min.js"></script>
+    <script type="text/javascript" src="/js/common.js"></script>
+    <script type="text/javascript" src="/fastpurge/fastpurge-history.js"></script>
   </body>
 </html>
