+++ popup.html
@@ -266,9 +266,10 @@
     <script type="text/javascript" src="js/jquery-3.1.1.min.js"></script>
     <script type="text/javascript" src="js/materialize.min.js"></script>
     <script type="text/javascript" src="js/common.js"></script>
-    <script type="text/javascript" src="cmanager/cmanager.js"></script>
-    <script type="text/javascript" src="js/purge.js"></script>
     <script type="text/javascript" src="js/popup.js"></script>
-    <script type="text/javascript" src="js/debugurls.js"></script>
+    <script type="text/javascript" src="cmanager/cmanager.js"></script>
+    <script type="text/javascript" src="fastpurge/fastpurge.js"></script>
+    <script type="text/javascript" src="debugreq/debugreq.js"></script>
+    <script type="text/javascript" src="piez/piez.js"></script>
   </body>
 </html>
